search refactor: move MainQueryParser and factory to parser package
