+++ kifi-backend/modules/search/app/com/keepit/search/MainSearcherFactory.scala
@@ -18,6 +18,8 @@ import com.keepit.common.akka.MonitoredAwait
 import com.keepit.common.akka.SafeFuture
 import com.keepit.search.user.UserIndexer
 import com.keepit.search.user.UserSearcher
+import com.keepit.search.query.parser.MainQueryParserFactory
+
 import scala.concurrent._
 import scala.concurrent.duration._
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
